// fifa-front/src/app/interfaces/player.interface.ts
export interface PlayerInterface {
  id?: number;
  fifa_version: string;
  fifa_update: string;
  player_face_url: string;
  long_name: string;
  player_positions: string;
  club_name?: string | null;
  nationality_name?: string | null;
  overall: number;
  potential: number;
  value_eur?: number | null;
  wage_eur?: number | null;
  age: number;
  height_cm?: number | null;
  weight_kg?: number | null;
  preferred_foot?: string | null;
  weak_foot?: number | null;
  skill_moves?: number | null;
  international_reputation?: number | null;
  work_rate?: string | null;
  body_type?: string | null;
  pace?: number | null;
  shooting?: number | null;
  passing?: number | null;
  dribbling?: number | null;
  defending?: number | null;
  physic?: number | null;
  attacking_crossing?: number | null;
  attacking_finishing?: number | null;
  attacking_heading_accuracy?: number | null;
  attacking_short_passing?: number | null;
  attacking_volleys?: number | null;
  skill_dribbling?: number | null;
  skill_curve?: number | null;
  skill_fk_accuracy?: number | null;
  skill_long_passing?: number | null;
  skill_ball_control?: number | null;
  movement_acceleration?: number | null;
  movement_sprint_speed?: number | null;
  movement_agility?: number | null;
  movement_reactions?: number | null;
  movement_balance?: number | null;
  power_shot_power?: number | null;
  power_jumping?: number | null;
  power_stamina?: number | null;
  power_strength?: number | null;
  power_long_shots?: number | null;
  mentality_aggression?: number | null;
  mentality_interceptions?: number | null;
  mentality_positioning?: number | null;
  mentality_vision?: number | null;
  mentality_penalties?: number | null;
  mentality_composure?: number | null;
  defending_marking?: number | null;
  defending_standing_tackle?: number | null;
  defending_sliding_tackle?: number | null;
  goalkeeping_diving?: number | null;
  goalkeeping_handling?: number | null;
  goalkeeping_kicking?: number | null;
  goalkeeping_positioning?: number | null;
  goalkeeping_reflexes?: number | null;
  goalkeeping_speed?: number | null;
  player_traits?: string | null;
}